Rearrange methods
- getKnitParams is only used by knitmanager, so should be a private method on it
